python/colab/requirements.txt
google-api-python-client
oauth2client
absl-py
